Refactor MemedBattle and related contracts to enhance functionality and maintainability. This update includes the introduction of a battle resolver interface, modifications to battle allocation and resolution logic, and updates to event emissions. Additionally, the battle ID type has been changed from uint128 to uint256 for improved scalability. The MemedTokenSale contract has been removed, and deployment scripts have been adjusted to accommodate these changes.

+++ contracts/MemedWarriorNFT.sol
@@ -124,7 +124,7 @@ contract MemedWarriorNFT is ERC721, Ownable, ReentrancyGuard {
     /** Get back the warrior NFTs to the user if they win the battle
      * @dev Get back the warrior NFTs to the user if they win the battle
      */
-    function getBackWarrior(uint128 _battleId) external {
+    function getBackWarrior(uint256 _battleId) external {
         Battle memory battle = memedBattle.getBattle(_battleId);
         require(battle.winner == memedToken, "Not the winner");
         UserBattleAllocation memory allocation = memedBattle.getBattleAllocations(_battleId, msg.sender, memedToken);
